Split CLI tests into two files
For task functionality and everything else.
