+++ tests/python/cli/test_cli_tasks.py
@@ -2,49 +2,22 @@
 #
 # SPDX-License-Identifier: MIT
 
-import io
 import json
 import os
 from pathlib import Path
-from typing import Optional
 
-import packaging.version as pv
 import pytest
-from cvat_sdk import Client, make_client
-from cvat_sdk.api_client import exceptions, models
+from cvat_sdk.api_client import exceptions
 from cvat_sdk.core.proxies.tasks import ResourceType, Task
 from PIL import Image
 
 from sdk.util import generate_coco_json
-from shared.utils.config import BASE_URL, USER_PASS
 from shared.utils.helpers import generate_image_file
 
-from .util import generate_images, https_reverse_proxy, run_cli
+from .util import TestCliBase, generate_images
 
 
-class TestCLI:
-    @pytest.fixture(autouse=True)
-    def setup(
-        self,
-        restore_db_per_function,  # force fixture call order to allow DB setup
-        restore_redis_inmem_per_function,
-        restore_redis_ondisk_per_function,
-        fxt_stdout: io.StringIO,
-        tmp_path: Path,
-        admin_user: str,
-    ):
-        self.tmp_path = tmp_path
-        self.stdout = fxt_stdout
-        self.host, self.port = BASE_URL.rsplit(":", maxsplit=1)
-        self.user = admin_user
-        self.password = USER_PASS
-        self.client = make_client(
-            host=self.host, port=self.port, credentials=(self.user, self.password)
-        )
-        self.client.config.status_check_period = 0.01
-
-        yield
-
+class TestCliTasks(TestCliBase):
     @pytest.fixture
     def fxt_image_file(self):
         img_path = self.tmp_path / "img_0.png"
@@ -85,27 +58,6 @@ class TestCLI:
         )
 
         return task
-
-    def run_cli(
-        self, cmd: str, *args: str, expected_code: int = 0, organization: Optional[str] = None
-    ) -> str:
-        common_args = [
-            f"--auth={self.user}:{self.password}",
-            f"--server-host={self.host}",
-            f"--server-port={self.port}",
-        ]
-
-        if organization is not None:
-            common_args.append(f"--organization={organization}")
-
-        run_cli(
-            self,
-            *common_args,
-            cmd,
-            *args,
-            expected_code=expected_code,
-        )
-        return self.stdout.getvalue()
 
     def test_can_create_task_from_local_images(self):
         files = generate_images(self.tmp_path, 5)
@@ -231,81 +183,6 @@ class TestCLI:
         assert task_id
         assert task_id != fxt_new_task.id
         assert self.client.tasks.retrieve(task_id).size == fxt_new_task.size
-
-    def test_can_warn_on_mismatching_server_version(self, monkeypatch, caplog):
-        def mocked_version(_):
-            return pv.Version("0")
-
-        # We don't actually run a separate process in the tests here, so it works
-        monkeypatch.setattr(Client, "get_server_version", mocked_version)
-
-        self.run_cli("ls")
-
-        assert "Server version '0' is not compatible with SDK version" in caplog.text
-
-    @pytest.mark.parametrize("verify", [True, False])
-    def test_can_control_ssl_verification_with_arg(self, verify: bool):
-        with https_reverse_proxy() as proxy_url:
-            if verify:
-                insecure_args = []
-            else:
-                insecure_args = ["--insecure"]
-
-            run_cli(
-                self,
-                f"--auth={self.user}:{self.password}",
-                f"--server-host={proxy_url}",
-                *insecure_args,
-                "ls",
-                expected_code=1 if verify else 0,
-            )
-            stdout = self.stdout.getvalue()
-
-        if not verify:
-            for line in stdout.splitlines():
-                int(line)
-
-    def test_can_control_organization_context(self):
-        org = "cli-test-org"
-        self.client.organizations.create(models.OrganizationWriteRequest(org))
-
-        files = generate_images(self.tmp_path, 1)
-
-        stdout = self.run_cli(
-            "create",
-            "personal_task",
-            ResourceType.LOCAL.name,
-            *map(os.fspath, files),
-            "--labels=" + json.dumps([{"name": "person"}]),
-            "--completion_verification_period=0.01",
-            organization="",
-        )
-
-        personal_task_id = int(stdout.split()[-1])
-
-        stdout = self.run_cli(
-            "create",
-            "org_task",
-            ResourceType.LOCAL.name,
-            *map(os.fspath, files),
-            "--labels=" + json.dumps([{"name": "person"}]),
-            "--completion_verification_period=0.01",
-            organization=org,
-        )
-
-        org_task_id = int(stdout.split()[-1])
-
-        personal_task_ids = list(map(int, self.run_cli("ls", organization="").split()))
-        assert personal_task_id in personal_task_ids
-        assert org_task_id not in personal_task_ids
-
-        org_task_ids = list(map(int, self.run_cli("ls", organization=org).split()))
-        assert personal_task_id not in org_task_ids
-        assert org_task_id in org_task_ids
-
-        all_task_ids = list(map(int, self.run_cli("ls").split()))
-        assert personal_task_id in all_task_ids
-        assert org_task_id in all_task_ids
 
     def test_auto_annotate_with_module(self, fxt_new_task: Task):
         annotations = fxt_new_task.get_annotations()
